:recycle: Refactor auth endpoints to share single directus client
